requirements.txt
importlib_metadata==3.1.0
tqdm==4.42.1
torch
opencv_python==4.5.1.48
matplotlib==3.5.1
plotext==3.1.3
torchvision
numpy==1.19.3
h5py==3.3.0
coloredlogs==15.0.1
Pillow==9.0.1
scikit_learn==1.0.2
argparse==1.1
timm
monai
pandas
captum
grad-cam
